moviepy_poc/requirements.txt
moviepy==2.2.1
natsort
pillow
pydub